Move project list to marketplace tab and update screens
